+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to
 
 ### Added
 
+- E2E tests for edge validation in collaborative editor
+  [#3724](https://github.com/OpenFn/lightning/issues/3724)
+
 ### Changed
 
 ### Fixed